add levels (WIP)
add basic level functionality, still working on import; django is killing me
here, it's croaking on perfectly valid filenames (which have single quote in
them) and it's not obvious how to allow it.  I either have to figure it out
or replace/rename all the files which I'm not keen to do.