add entrypoint to collectstatic
need to put django/admin static files somewhere accessible so /admin can work

I should probably put them on a volume and serve from massassi-static (nginx)
instead of gunicorn.  But it's for admin only so maybe it will be ok?

fix holiday logo import script to import a direct copy from prod; before this
change it was importing duplicates for the older files