support file uploads
Uploading a file is a 3 step process:

stage the file (that creates a "blank" row in the db and returns a file_id)

upload the file (use the new .upload endpoint which requires the file_id)

set the remaining file info (title, version, description)

It's a bit of a pain but this allows the user to start uploading a file and
then fill out the rest of the file info as the file is uploading (rather than
wait for it to finish, then add the rest of the info).

massassi/models.py: On MassassiModelWithFile, files were always supposed to
be optional, but the way the save() method was written it would error out
if file wasn't defined; guard against that.