fix levels "sort by"
Sort by form only worked on the first page, once you click to a different
page the sort you chose would be lost.  This hopefully fixes that.

Put the sortby and number of results per page into the template context
so it can be persisted by the templates.

Update the fancy_pager template to look for those context variables and
add them to all the pager links.